backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
langchain>=0.1.0
langgraph>=0.0.20
langchain-groq>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sse-starlette>=1.8.0
